test: wait for latest process definition version before opening Processes tab
Backport of #58102 to stable/8.9. Moves the indexing-lag helper from
the spec file into utils/zeebeClient.ts, using the existing SDK client
(searchProcessDefinitions with isLatestVersion) and the shared sleep
helper instead of a hand-rolled fetch with manual auth.

Verified locally against camunda/camunda:8.9-SNAPSHOT (full
processes-page.spec.ts green, 10 passed, 1 skipped pre-existing,
0 flaky).

relates to #40045 #58050
